Move download functions from fileUtil to netUtil
